remove campaign partial update

+++ apps/campaign/api/serializers.py
@@ -1,6 +1,6 @@
 import datetime
 
-from rest_framework import serializers
+from rest_framework import serializers, exceptions
 
 from apps.campaign.models import Province, Campaign, CampaignContent, CampaignSchedule, TargetDevice
 
@@ -44,13 +44,28 @@ class CampaignSerializer(serializers.ModelSerializer):
         fields = '__all__'
         read_only_fields = ['owner', 'status']
 
+    def get_extra_kwargs(self):
+        extra_kwargs = super().get_extra_kwargs()
+        action = self.context['view'].action
+
+        if action in ['update']:
+            kwargs = extra_kwargs.get('medium', {})
+            kwargs['read_only'] = True
+            extra_kwargs['medium'] = kwargs
+
+        return extra_kwargs
+
     def validate(self, attrs):
 
-        medium = attrs.get('medium')
+        if self.instance:
+            medium = self.instance.medium
+        else:
+            medium = attrs.get('medium')
+
         publishers = attrs.get('publishers', [])
         categories = attrs.get('categories', [])
 
-        if publishers == categories is None:
+        if len(publishers) == 0 and len(categories) == 0:
             raise serializers.ValidationError("Publishers and categories both can not be empty")
 
         for category in categories:
@@ -61,7 +76,7 @@ class CampaignSerializer(serializers.ModelSerializer):
             if publisher.medium != medium:
                 raise serializers.ValidationError("Campaign's medium and publisher's medium must be the same.")
 
-        schedule_set = attrs.get('campaignschedule_set', [])[:]
+        schedule_set = attrs.get('campaignschedule_set')[:]
         while schedule_set:
             schedule = schedule_set.pop()
             for i in schedule_set:
@@ -76,8 +91,8 @@ class CampaignSerializer(serializers.ModelSerializer):
         return attrs
 
     def create(self, validated_data):
-        targetdevice_set = validated_data.pop("targetdevice_set", [])
-        schedule_set = validated_data.pop("campaignschedule_set", [])
+        targetdevice_set = validated_data.pop("targetdevice_set")
+        schedule_set = validated_data.pop("campaignschedule_set")
 
         if validated_data['start_date'] is None:
             validated_data["start_date"] = datetime.date.today()
@@ -93,37 +108,42 @@ class CampaignSerializer(serializers.ModelSerializer):
         return campaign
 
     def update(self, instance, validated_data):
-        targetdevice_set = validated_data.pop("targetdevice_set")
-        schedule_set = validated_data.pop("campaignschedule_set")
+        targetdevice_set = validated_data.pop("targetdevice_set", None)
+        schedule_set = validated_data.pop("campaignschedule_set", None)
 
         instance = super().update(instance, validated_data)
 
-        targetdevice_id_list = [i.get("id") for i in targetdevice_set if i.get("id")]
-        TargetDevice.objects.filter(campaign=instance).exclude(id__in=targetdevice_id_list).delete()
+        if targetdevice_set is not None:
+            targetdevice_id_list = [i.get("id") for i in targetdevice_set if i.get("id")]
+            TargetDevice.objects.filter(campaign=instance).exclude(id__in=targetdevice_id_list).delete()
 
-        schedule_id_list = [i.get("id") for i in schedule_set if i.get("id")]
-        CampaignSchedule.objects.filter(campaign=instance).exclude(id__in=schedule_id_list).delete()
+            for targetdevice_data in targetdevice_set:
+                targetdevice_id = targetdevice_data.get('id', None)
+                if targetdevice_id:
+                    try:
+                        targetdevice = TargetDevice.objects.get(id=targetdevice_id, campaign=instance)
+                    except TargetDevice.DoesNotExist:
+                        raise exceptions.ParseError
+                    targetdevice.device = targetdevice_data.get('device', targetdevice.device)
+                    targetdevice.service_provider = targetdevice_data.get('service_provider',
+                                                                          targetdevice.service_provider)
+                    targetdevice.save()
+                else:
+                    TargetDevice.objects.create(campaign=instance, **targetdevice_data)
 
-        for targetdevice_data in targetdevice_set:
-            targetdevice_id = targetdevice_data.get('id', None)
-            if targetdevice_id:
-                targetdevice = TargetDevice.objects.get(id=targetdevice_id, campaign=instance)
-                targetdevice.device = targetdevice_data.get('device', targetdevice.device)
-                targetdevice.service_provider = targetdevice_data.get('service_provider', targetdevice.service_provider)
-                targetdevice.save()
-            else:
-                TargetDevice.objects.create(campaign=instance, **targetdevice_data)
-
-        for schedule_data in schedule_set:
-            schedule_id = schedule_data.get('id', None)
-            if schedule_id:
-                schedule = CampaignSchedule.objects.get(id=schedule_id, campaign=instance)
-                schedule.day = schedule_data.get('day', schedule.day)
-                schedule.start_time = schedule_data.get('start_time', schedule.start_time)
-                schedule.end_time = schedule_data.get('end_time', schedule.end_time)
-                schedule.save()
-            else:
-                CampaignSchedule.objects.create(campaign=instance, **schedule_data)
+        if schedule_set is not None:
+            schedule_id_list = [i.get("id") for i in schedule_set if i.get("id")]
+            CampaignSchedule.objects.filter(campaign=instance).exclude(id__in=schedule_id_list).delete()
+            for schedule_data in schedule_set:
+                schedule_id = schedule_data.get('id', None)
+                if schedule_id:
+                    schedule = CampaignSchedule.objects.get(id=schedule_id, campaign=instance)
+                    schedule.day = schedule_data.get('day', schedule.day)
+                    schedule.start_time = schedule_data.get('start_time', schedule.start_time)
+                    schedule.end_time = schedule_data.get('end_time', schedule.end_time)
+                    schedule.save()
+                else:
+                    CampaignSchedule.objects.create(campaign=instance, **schedule_data)
 
         return instance
